requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apprise==1.8.0
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
croniter==6.0.0
dotenv==0.9.9
fastapi==0.115.12
h11==0.14.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pathvalidate==3.2.3
pillow==11.2.1
platformdirs==4.3.8
PlexAPI==4.16.1
pluggy==1.5.0
prettytable==3.16.0
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qbittorrent-api==2024.3.60
ratelimit==2.2.1
requests==2.32.3
requests-oauthlib==2.0.0
ruamel.yaml.clib==0.2.8
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
Unidecode==1.3.8
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
